socket logic moved to its own module
