+++ server/serverbuild.js
@@ -1,0 +1,7 @@
+const express = require('express');
+
+const app = express();
+const http = require('http').createServer(app);
+const io = require('socket.io')(http);
+
+module.exports = { express, app, http, io };
